file_list_server/requirements.txt
Flask_SQLAlchemy==2.4.0
gunicorn==19.9.0
Flask==1.0.2
gevent==1.4.0
Flask_HTTPAuth==3.2.4
pycrypto==2.6.1
